% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_thVals2resTable.R
\name{add_thVals2resTable}
\alias{add_thVals2resTable}
\title{Add accuracies at specific thresholds to summary file}
\usage{
add_thVals2resTable(parcc = NULL, verbose = FALSE)
}
\arguments{
\item{parcc}{The experiments settings. If NULL they are 
derived with \code{get_settings()}.}

\item{verbose}{Print progress}
}
\value{
NULL, adds to summary file
}
\description{
Add accuracies at specific thresholds to summary file
}
\details{
This function loads (with \code{\link{get_summaryFile}}) 
the previously with  \code{\link{run_experiments}} calculated
results and adds the accuracies for the thresholds minTP_T, 
tenTP_T, mxSSSP_T, eqSSSP_T which are derived from the 
PU validation data.
}

