% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLine2csv.R
\name{addLine2csv}
\alias{addLine2csv}
\title{Add a result line to the csv file}
\usage{
addLine2csv(line, file)
}
\arguments{
\item{line}{the line to add}

\item{file}{the file to add to}
}
\description{
Add a result line to the csv file
}

