% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddp_perfLoss.R
\name{ddp_perfLoss}
\alias{ddp_perfLoss}
\title{Title}
\usage{
ddp_perfLoss(x, metric = "mxK", best = "star.k.k", others = NULL)
}
\arguments{
\item{x}{}

\item{metric}{}

\item{best}{A string giving the benchmarck, i.e. the 
accuracy from which to substract the accuracy realized with 
another approach. This must be a character combined by a 
value of x$ms and x$ts separated by a dot, e.g. the default 
\code{star.k.k.}.
the realized accuracy}

\item{others}{}
}
\description{
Title
}

