% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_settings}
\alias{get_settings}
\title{Get settings required for using package occCompare}
\usage{
get_settings()
}
\description{
It is assumed that the settings restored as
option \code{occc_sttings}, thus callable via 
\code{getOption("occc_settings")}.
}

