% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tasks.R
\name{get_tasks}
\alias{get_tasks}
\title{Get all tasks (of a method) in a data frame}
\usage{
get_tasks(method = NULL)
}
\arguments{
\item{method}{If NULL tasks are given for each method. Else specify the given only the tasks of a method.}
}
\value{
dta frame
}
\description{
Get all tasks (of a method) in a data frame
}

