% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_exp_dot_trainUtest.R
\name{get_tprUppp}
\alias{get_tprUppp}
\title{Title}
\usage{
get_tprUppp(ev, idx, prefix, sep = "_")
}
\arguments{
\item{ev}{}

\item{idx}{}

\item{prefix}{}

\item{sep}{}
}
\description{
Title
}

