% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_parameter_grids.R
\name{ggp_parameter_grids}
\alias{ggp_parameter_grids}
\title{Title}
\usage{
ggp_parameter_grids(method, x, y, facet_x = NULL, res = NULL,
  tasks = NULL, grid_boder_cases_only = TRUE, optcrit = "mxK",
  ignore = NULL, print_single_grids = FALSE, verbose = FALSE)
}
\arguments{
\item{method}{}

\item{x}{}

\item{y}{}

\item{facet_x}{}

\item{res}{}

\item{tasks}{}

\item{grid_boder_cases_only}{}

\item{optcrit}{}

\item{ignore}{}

\item{verbose}{}
}
\description{
Title
}

