% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_perfloss.R
\name{ggp_perfloss}
\alias{ggp_perfloss}
\title{Title}
\usage{
ggp_perfloss(x, ignore_methods = "star", factorVals4plot = get_fVals4plot(),
  cols4plot = get_cols4plot(), box_lwd = 0.2, box_outlier_size = 0.2,
  axis.text.size = 10, strip.text.size = 10, rotate_axis.title.y = TRUE,
  panel.margin.y = 0, panel.margin.x = 0, hlines = c(0, 0.1))
}
\arguments{
\item{x}{}

\item{ignore_methods}{}

\item{factorVals4plot}{}

\item{cols4plot}{}

\item{box_lwd}{}

\item{box_outlier_size}{}

\item{axis.text.size}{}

\item{strip.text.size}{}

\item{rotate_axis.title.y}{}

\item{panel.margin.y}{}

\item{panel.margin.x}{}
}
\description{
Title
}

