% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_results.R
\name{load_results}
\alias{load_results}
\title{Load results written by \code{run_experiments}.}
\usage{
load_results(what, bdir = NULL, fullResults = T, maxentDef = TRUE,
  verbosity = 1, file = NULL, overwrite = FALSE)
}
\arguments{
\item{bdir}{The base directory with the summary tables written by run_experiments/run_exp.}

\item{fullResults}{if \code{TRUE} the full results table is returned, 
i.e. all eperiments/models defined in \code{get_settings()} even if they are not 
found in the results stored on disk (=> \code{NA}).}
}
\description{
...
}

