% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection}
\alias{model_selection}
\title{Title}
\usage{
model_selection(x, ms = NULL, acc_metric = "mxK", sort_ps_first = TRUE,
  verbose = FALSE, file = NULL, overwrite = FALSE)
}
\arguments{
\item{x}{list with the results returned by \code{load_results}}

\item{ms}{data frame with the columns 
\code{ps}, \code{ts} and \code{cl}.}

\item{acc_metric}{Defaults: mxK. Accuracy metric.}
}
\description{
Title
}

