% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_exp.R
\name{run_exp}
\alias{run_exp}
\title{Title}
\usage{
run_exp(rs, looppar, tuneGrid, fun = ".trainUtest", doPar = TRUE,
  loadOnly = FALSE, overwrite = FALSE, rm_ifNotPuResampl = NULL)
}
\arguments{
\item{looppar}{}

\item{tuneGrid}{}

\item{fun}{}

\item{doPar}{}

\item{loadOnly}{}

\item{overwrite}{}

\item{rm_ifNotPuResampl}{}
}
\description{
Title
}

