% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiments.R
\name{run_experiments}
\alias{run_experiments}
\title{run_experiments}
\usage{
run_experiments(parcc = NULL, fun = ".trainUtest", doPar = TRUE,
  debug = FALSE, overwrite = FALSE, overwrite_resTable = FALSE,
  loadOnly = FALSE, ...)
}
\arguments{
\item{doPar}{If \code{TRUE} and if possible uses parallel processing.}

\item{debug}{...}

\item{overwrite}{Set to \code{TRUE} if all the already available results should be overwritten.}

\item{overwrite_resTable}{Delete the resTables before collecting the results.}

\item{loadOnly}{...}
}
\value{
\code{NULL}
}
\description{
Run the experiments and save the data to disk.
}
\details{
\strong{\code{parcc}} must be a list, 
which (at least) contains the following elements: \cr
\describe{
 \item{\code{dn.res}}{A directory in which to write the raw results.}
 \item{\code{seed}}{Vector with seed points 
 (single element passed to \code{get_refset})}
 \item{\code{nP}}{Vector with the number of positive samples 
 (single element passed to \code{get_refset})}
 \item{\code{nU}}{Vector with the number of unlabeled samples 
 (single element passed to \code{get_refset})}
 \item{\code{fset}}{Vector with the predictor/feature set names 
 (single element passed to \code{get_refset})}
 \item{\code{scaling}}{Vector with scaling type name, currently 
 only \code{ccs01}}
 \item{\code{method}}{Vector of OCC method names that can be 
 passed to \code{\link[oneClass]{trainOcc}}.}
 \item{\code{resampling}}{Vector of resampling types. (So far: 
 \code{val}, \code{cv10}, \code{cv10_pu})}
 \item{\code{param}}{Named list (names as \code{parcc$method}) 
 each of which containing the parameters for the respective method.}
}
\strong{\code{get_refset}} must be a function, which 
takes the arguments
\describe{
 \item{\code{seed}}{Any element of \code{parcc$seed}}
 \item{\code{nP}}{Any element of \code{parcc$nP}}
 \item{\code{nU}}{Any element of \code{parcc$nU}}
 \item{\code{fset}}{Any element of \code{parcc$fset}}
 \item{\code{args.rs}}{Optional, but usually a list which is available as 
 \code{args.rs} in the user defined function \code{get_refset()}}
 } 
The function must returns a \code{data.frame} with the columns
\describe{
 \item{\code{tr}}{logical, TRUE for training pixels}
 \item{\code{y}}{the class labels \{0, 1, ..., nClasses\}, with 0: unlabeled)}
 \item{\code{...}}{feature columns at positions 3:(nFeatures+2) with any names.}
 }
}

